+++ views/partial/shop_nav.hbs
@@ -37,12 +37,12 @@
       </div>
       <div class="flexed-item nav-end">
         {{#unless isLoggedIn}}
-        <a  href="/signin" id="nav_login" class="button-ternary flexed-item nav-end-item">
+        <a  href="/auth/signin" id="nav_login" class="button-ternary flexed-item nav-end-item">
           <span class="btn-text">Sign in</span>
         </a>
         {{/unless}}
         {{#if isLoggedIn}}
-        <a  href="/logout" id="nav_logout" class="button-ternary flexed-item nav-end-item">
+        <a  href="/auth/logout" id="nav_logout" class="button-ternary flexed-item nav-end-item">
           <span class="btn-text">Log out</span>
         </a>
         {{/if}}
