+++ views/products_add.hbs
@@ -1,7 +1,7 @@
 <div class="big-container product-manage-wrapper flexed-item">
-  <form action="/dashboard_sneaker/prod-add" method="post" class="add-product-form" enctype="multiparty/form-data">
+  <form action="/dashboard_sneaker/prod-add" method="post" class="add-product-form" enctype="multipart/form-data">
     {{#if msg}}
-    <p class="msg {{msg.statu}}">{{msg.txt}}</p>
+    <p class="msg {{msg.status}}">{{msg.txt}}</p>
     {{/if}}
     <h3 class="form-title">Add sneakers to the collection</h3>
     <div class="form-item-wrapper flexed-item">
