+++ views/signin.hbs
@@ -6,7 +6,7 @@
   {{#if msg}}
   <p class="message {{msg.status}}">{{msg.text}}</p>
   {{/if}}
-  <form method="post" action="/signin" id="form_user">
+  <form method="post" action="/auth/signin" id="form_user">
     <div class="form-item-wrapper flexed-item">
       <label for="email" class="label-ref label">Email</label>
       <input id="email" type="email" name="email" class="input-text input" placeholder="albert.biceps@gmail.com" required value="yo@gui.com">
@@ -17,5 +17,5 @@
     </div>
     <button class="button-primary white">Log in !</button>
   </form>
-  <p class="form-sub-infos">Don't have an account yet ? <a href="/signup" class="button-link-form">Sign Up</a></p>
+  <p class="form-sub-infos">Don't have an account yet ? <a href="/auth/signup" class="button-link-form">Sign Up</a></p>
 </section>
