+++ views/signup.hbs
@@ -6,7 +6,7 @@
   {{#if msg}}
   <p class="message {{msg.status}}">{{msg.text}}</p>
   {{/if}}
-  <form method="post" action="/signup" id="form_user">
+  <form method="post" action="/auth/signup" id="form_user">
     <div class="form-item-wrapper flexed-item">
       <label for="firstname" class="label-ref label">Name</label>
       <input id="firstname" type="text" name="name" class="input-text input" placeholder="Albert" required value="yo">
@@ -25,5 +25,5 @@
     </div>
     <button class="button-primary white">Create Account</button>
   </form>
-  <p class="form-sub-infos">Already have an account ? <a href="/signin" class="button-link-form">Sign In</a></p>
+  <p class="form-sub-infos">Already have an account ? <a href="/auth/signin" class="button-link-form">Sign In</a></p>
 </section>
